+++ src/PortalBase/Exploration/ExplorerStack.php
@@ -42,47 +42,8 @@ class ExplorerStack implements ExplorerStackInterface, LoggerAwareInterface
             return [];
         }
 
-        $this->logger->debug(\sprintf('Execute FlowComponent explorer: %s', $this->getOrigin($explorer)));
+        $this->logger->debug(\sprintf('Execute FlowComponent explorer: %s', \get_class($explorer)));
 
         return $explorer->explore($context, $this);
-    }
-
-    public function listOrigins(): array
-    {
-        $origins = [];
-        foreach ($this->explorers as $explorer) {
-            $origins[] = $this->getOrigin($explorer);
-        }
-
-        return $origins;
-    }
-
-    protected function getOrigin(ExplorerContract $explorer): string
-    {
-        if ($explorer instanceof ShorthandExplorer) {
-            $runMethod = $explorer->getRunMethod();
-
-            if ($runMethod instanceof \Closure) {
-                $reflection = new \ReflectionFunction($runMethod);
-
-                return $reflection->getFileName() . '::run:' . $reflection->getStartLine();
-            }
-
-            $isAllowedMethod = $explorer->getIsAllowedMethod();
-
-            if ($isAllowedMethod instanceof \Closure) {
-                $reflection = new \ReflectionFunction($isAllowedMethod);
-
-                return $reflection->getFileName() . '::isAllowed:' . $reflection->getStartLine();
-            }
-
-            $this->logger->warning('ExplorerStack contains unconfigured short-notation explorer', [
-                'code' => 1637421327,
-            ]);
-        }
-
-        $reflection = new \ReflectionClass($explorer);
-
-        return $reflection->getFileName() . ':' . $reflection->getStartLine();
     }
 }